move octree.rs into module folder
