+++ src/octree/mod.rs
@@ -13,8 +13,11 @@ use crate::{
     math::{min_and_max, sort_by_indices},
     mesh, morton,
     sources::{h_current_point, h_current_tet4},
-    types::Vec3,
+    types::{Vec3, Mat3}
 };
+
+mod bbox;
+use bbox::BoundingBox;
 
 use std::{f64::consts::PI, thread, time::Instant};
 
@@ -32,84 +35,6 @@ const INVALID_NODE: u32 = u32::MAX;
 /// node and the level in the tree
 pub fn size_at_level(side_length: f64, level: u8) -> f64 {
     side_length / (2f64.powi(level as i32))
-}
-
-/// Determines the location and extent of a collection of source points
-#[derive(Debug, Clone, Copy)]
-pub struct BoundingBox {
-    xc: f64,
-    yc: f64,
-    zc: f64,
-    pub side_length: f64,
-    xbounds: (f64, f64),
-    ybounds: (f64, f64),
-    zbounds: (f64, f64),
-}
-
-// Get the maximum side length of a bounding box cube that encloses all x,y,z bounds
-fn side_length_from_bounds(xbounds: (f64, f64), ybounds: (f64, f64), zbounds: (f64, f64)) -> f64 {
-    let xrange: f64 = xbounds.1 - xbounds.0;
-    let yrange: f64 = ybounds.1 - ybounds.0;
-    let zrange: f64 = zbounds.1 - zbounds.0;
-
-    let xymax = match xrange > yrange {
-        true => xrange,
-        false => yrange,
-    };
-
-    match xymax > zrange {
-        true => xymax,
-        false => zrange,
-    }
-}
-
-impl BoundingBox {
-    pub fn from_centroids(centroids: (&[f64], &[f64], &[f64])) -> Option<Self> {
-        // TODO: length check
-
-        let xbounds: Option<(f64, f64)> = min_and_max(centroids.0);
-        let ybounds: Option<(f64, f64)> = min_and_max(centroids.1);
-        let zbounds: Option<(f64, f64)> = min_and_max(centroids.2);
-
-        if xbounds.is_none() | ybounds.is_none() | zbounds.is_none() {
-            return None;
-        }
-
-        let (xb, yb, zb) = (xbounds.unwrap(), ybounds.unwrap(), zbounds.unwrap());
-        let side_length = side_length_from_bounds(xb, yb, zb);
-        let xc: f64 = xb.0 + 0.5 * side_length;
-        let yc: f64 = yb.0 + 0.5 * side_length;
-        let zc: f64 = zb.0 + 0.5 * side_length;
-
-        Some(Self {
-            xc,
-            yc,
-            zc,
-            side_length,
-            xbounds: xb,
-            ybounds: yb,
-            zbounds: zb,
-        })
-    }
-
-    /// TODO: fix this so there's no data copy
-    pub fn from_centroids_vec(centroids: &[Vec3]) -> Self {
-        let n: usize = centroids.len();
-        let mut x: Vec<f64> = vec![0.0; n];
-        let mut y: Vec<f64> = vec![0.0; n];
-        let mut z: Vec<f64> = vec![0.0; n];
-        for i in 0..n {
-            x[i] = centroids[i][0];
-            y[i] = centroids[i][1];
-            z[i] = centroids[i][2];
-        }
-
-        Self::from_centroids((&x, &y, &z)).unwrap()
-    }
-
-    pub fn min_corner(&self) -> (f64, f64, f64) {
-        (self.xbounds.0, self.ybounds.0, self.zbounds.0)
-    }
 }
 
 /// Return the morton code prefix at a given level of the tree
@@ -262,6 +187,13 @@ impl InteractionList {
     }
 }
 
+// Aggregated moments at each level of the octree 
+#[derive(Debug)]
+struct NodeMoments {
+    monopole: Vec<Vec3>,
+    dipole: Vec<Mat3>
+}
+
 /// An octree constructed from Nsources tet4 elements
 #[derive(Debug)]
 pub struct Octree {
@@ -272,8 +204,8 @@ pub struct Octree {
     // Data about the octree itself (node-level data)
     bbox: BoundingBox,
     topology: Topology,
-    j_moments: Option<Vec<Vec3>>,
-    m_moments: Option<Vec<Vec3>>,
+    j_moments: Option<NodeMoments>,
+    m_moments: Option<NodeMoments>,
 
     // Physics and geometry data about the sources in the octree
     sources: Sources,
@@ -290,20 +222,11 @@ impl Octree {
     ) -> Self {
         let max_depth: u8 = 21;
 
-        let sort_timer = Instant::now();
         let (codes, idx_sorted, bbox, sources) =
             Self::sort_sources(nodes, connectivity, jdensity, mvectors, max_depth);
-        // println!(
-        //     "Source sort time: {:.3} sec",
-        //     sort_timer.elapsed().as_secs_f64()
-        // );
 
-        let top_timer = Instant::now();
         let topology = Self::build_topology(&sources, &codes, &bbox, max_depth, leaf_threshold);
-        // println!(
-        //     "Topology time: {:.3} sec",
-        //     top_timer.elapsed().as_secs_f64()
-        // );
+
 
         let mut octree: Octree = Octree {
             codes,
